move helper methods to util module
