+++ operators/src/source/mod.rs
@@ -6,8 +6,8 @@ pub use self::csv::{
     CsvGeometrySpecification, CsvSource, CsvSourceParameters, CsvSourceStream, CsvTimeSpecification,
 };
 pub use self::gdal_source::{
-    add_ndvi_data_set, create_ndvi_meta_data, GdalDataSetParameters, GdalLoadingInfo,
-    GdalMetaDataRegular, GdalSource, GdalSourceParameters, GdalSourceProcessor,
+    FileNotFoundHandling, GdalDataSetParameters, GdalLoadingInfo, GdalMetaDataRegular, GdalSource,
+    GdalSourceParameters, GdalSourceProcessor,
 };
 pub use self::ogr_source::{
     OgrSource, OgrSourceColumnSpec, OgrSourceDataset, OgrSourceDatasetTimeType, OgrSourceErrorSpec,
