+++ operators/src/util/mod.rs
@@ -1,3 +1,4 @@
+pub mod gdal;
 pub mod input;
 pub mod number_statistics;
